fix(INV2-15): align Product.id type with INTEGER schema across the codebase (#7)

+++ src/main/java/com/inventra/api/core/service/stockbatch/model/request/RegisterStockEntryRequest.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record RegisterStockEntryRequest(
-        Long productId,
+        Integer productId,
         Integer kitchenId,
         Integer supplierId,
         String batchNumber,
